backend/budgeteer_mcp_agent/requirements.txt
modelcontextprotocol>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.18.0
httpx>=0.24.0
mcp>=1.13.0
